Fixed : header problem

+++ src/ImageProcessing/Rotation.h
@@ -3,6 +3,6 @@
 
 #include "SDL2/SDL.h"
 
-SDL_Surface *rotation(SDL_Surface *image, int angle);
-SDL_Surface *Rotation_shearing(SDL_Surface *image, int angle);
+SDL_Surface *rotation(SDL_Surface *image, double angle);
+SDL_Surface *Rotation_shearing(SDL_Surface *image, double angle);
 #endif
